@@ -7,7 +7,7 @@ name = "conventional-commits-generator"
 version = "2.2.5"
 description = "Interactive CLI tool for creating standardized commit messages following conventional commits"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = {file = "LICENSE"}
 authors = [
     {name = "Egydio Bolonhezi", email = "egydiobolonhezi@gmail.com"}
@@ -21,7 +21,6 @@ classifiers = [
     "Natural Language :: English",
     "Operating System :: OS Independent",
     "Programming Language :: Python :: 3",
-    "Programming Language :: Python :: 3.8",
     "Programming Language :: Python :: 3.9",
     "Programming Language :: Python :: 3.10",
     "Programming Language :: Python :: 3.11",
@@ -37,7 +36,14 @@ dependencies = [
 [project.optional-dependencies]
 dev = [
     "pre-commit >= 3.5.0",
-    "prompt_toolkit >= 3.0.20"
+    "prompt_toolkit >= 3.0.20",
+    "tox >= 4.0.0"
+]
+test = [
+    "pytest >= 7.4.0",
+    "pytest-cov >= 4.1.0",
+    "pytest-mock >= 3.11.0",
+    "hypothesis >= 6.82.0"
 ]
 
 [project.urls]
